+++ belle-sip/src/belle_sip_headers_impl.c
@@ -632,7 +632,7 @@ int belle_sip_header_user_agent_get_products_as_string(const belle_sip_header_us
 	}
 	if (result>0) value[result-1]='\0'; /*remove last space */
 
-	return result-1;
+	return (int)result-1;
 }
 
 /**************************
@@ -964,7 +964,7 @@ BELLE_SIP_PARSE(header_service_route)
 */
 struct _belle_sip_header_content_length  {
 	belle_sip_header_t header;
-	unsigned int content_length;
+	size_t content_length;
 };
 
 static void belle_sip_header_content_length_destroy(belle_sip_header_content_length_t* content_length) {
@@ -985,8 +985,8 @@ belle_sip_error_code belle_sip_header_content_length_marshal(belle_sip_header_co
 
 BELLE_SIP_NEW_HEADER(header_content_length,header,BELLE_SIP_CONTENT_LENGTH)
 BELLE_SIP_PARSE(header_content_length)
-GET_SET_INT(belle_sip_header_content_length,content_length,unsigned int)
-belle_sip_header_content_length_t* belle_sip_header_content_length_create (int content_length)  {
+GET_SET_INT(belle_sip_header_content_length,content_length,size_t)
+belle_sip_header_content_length_t* belle_sip_header_content_length_create (size_t content_length)  {
 	belle_sip_header_content_length_t* obj;
 	obj = belle_sip_header_content_length_new();
 	belle_sip_header_content_length_set_content_length(obj,content_length);
